introduces `utils.structs`
- add test for LazyOption init
- move ZnTypes and DVCOptions to structs module
