+++ tests/unit_tests/utlis/test_structs.py
@@ -1,0 +1,8 @@
+import pytest
+
+from zntrack.utils.structs import LazyOption
+
+
+def test_noLazyOptionInit():
+    with pytest.raises(ValueError):
+        LazyOption()
